% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{base_nw.adol}
\alias{base_nw.adol}
\title{Construct Base Network for Model Estimation and Simulation}
\usage{
base_nw.adol(nwstats)
}
\arguments{
\item{nwstats}{An object of class \code{nwstats}, as output from
\code{\link{calc_nwstats.adol}}.}
}
\description{
Initializes the base network for model estimation within
             \code{netest}.
}
\details{
This function takes the output of \code{\link{calc_nwstats.adol}} and constructs
an empty network with the necessary attributes for race, square root of age based on entry probabilities is supplied,
sexual role class, risk catagory and a starting distribution for sexual debut if debut probabilities are supplied. 
This base network is used for all three network
estimations.
}
\seealso{
The final vertex attributes on the network for cross-network degree are
calculated and set on the network with \code{\link{assign_degree}}.
}

