% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.debut.R
\name{debut.adol}
\alias{debut.adol}
\title{Sexual Debut Module}
\usage{
debut.adol(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns \code{dat} after updating the nodal attribute
\code{has.debuted}. The \code{has.debuted} vertex attribute is also
updated on the three networks.
}
\description{
Module for sexual debut over time for active nodes in the population.  Having debuted is a condition for AI but it does not
require that AI occur.  It is a debut in terms of identity and willingness not action.
}
\keyword{module}

