% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{init.adol}
\alias{init.adol}
\title{Epidemic Model Initial Conditions for MARDHAM Models}
\usage{
init.adol(nwstats, prev.B = 0.15, prev.W = 0.15, ...)
}
\arguments{
\item{nwstats}{Target statistics for the network model. An object of class
\code{nwstats} output from \code{\link{calc_nwstats.adol}}.}

\item{prev.B}{Initial disease prevalence among black MSM.}

\item{prev.W}{Initial disease prevalence among white MSM.}

\item{...}{Additional arguments passed to function.}

\item{birth.age.prob}{Vector of length ages indicating the probability of entering the population at age.}

\item{debut.entry.prob}{probability of sexually debuing at entry.}

\item{debut.prob}{probability per timestep of debuing given not yet debuted.}

\item{birth.age.min}{the minimum age of a new entrant.}
}
\value{
A list object of class \code{init.adol}, which can be passed to EpiModel
function \code{\link{netsim}}.
}
\description{
Sets the initial conditions for a stochastic epidemic models
             simulated with \code{\link{netsim}}.
}

