% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.riskhist.adol.R
\name{riskhist.adol}
\alias{riskhist.adol}
\title{Risk History Module for Adolecents}
\usage{
riskhist.adol(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}

\item{uaicount}{is the number of UAI in the past 6 months.}
}
\description{
Module function to track the risk history defined by UAI of uninfected persons
             for purpose of intervention targeting.
}
\keyword{module}

